insurance_copilot/requirements.txt
langchain
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
pypdf
python-dotenv
pydantic
